Move impl to header
